Fix MSVC compiler warning
reg.: 'return': conversion from 'size_t' to 'int', possible loss of data (C4267)
Introduced with 1668e1fce9

+++ src/engraving/rendering/score/layoutcontext.cpp
@@ -151,7 +151,7 @@ const std::vector<Part*>& DomAccessor::parts() const
     return score()->parts();
 }
 
-int DomAccessor::visiblePartCount() const
+size_t DomAccessor::visiblePartCount() const
 {
     IF_ASSERT_FAILED(score()) {
         return 0;
